backend/requirements.txt
flask
uvicorn[standard]
pillow